splits gc-core in two since File garbage_collector_core.py has 508 lines of code (exceeds 500)
