+++ services/web/server/src/simcore_service_webserver/garbage_collector_core.py
@@ -8,17 +8,19 @@ from typing import Any, Dict, List, Optional, Set, Tuple
 
 import asyncpg.exceptions
 from aiohttp import web
-from aiopg.sa.result import RowProxy
 from aioredlock import Aioredlock
 from servicelib.utils import logged_gather
 from simcore_postgres_database.errors import DatabaseError
 from simcore_postgres_database.models.users import UserRole
 
 from . import director_v2_api, users_exceptions
-from .db_models import GroupType
 from .director.director_exceptions import DirectorException, ServiceNotFoundError
 from .garbage_collector_settings import GUEST_USER_RC_LOCK_FORMAT
-from .groups_api import get_group_from_gid
+from .garbage_collector_utils import (
+    get_new_project_owner_gid,
+    remove_user,
+    replace_current_owner,
+)
 from .projects.projects_api import (
     delete_project,
     get_project_for_user,
@@ -26,19 +28,12 @@ from .projects.projects_api import (
     is_node_id_present_in_any_project_workbench,
     remove_project_interactive_services,
 )
-from .projects.projects_db import APP_PROJECT_DBAPI, ProjectAccessRights
+from .projects.projects_db import APP_PROJECT_DBAPI
 from .projects.projects_exceptions import ProjectNotFoundError
 from .redis import get_redis_lock_manager
 from .resource_manager.registry import RedisResourceRegistry, get_registry
-from .users_api import (
-    delete_user,
-    get_guest_user_ids_and_names,
-    get_user,
-    get_user_id_from_gid,
-    get_user_role,
-)
+from .users_api import get_guest_user_ids_and_names, get_user, get_user_role
 from .users_exceptions import UserNotFoundError
-from .users_to_groups_api import get_users_for_gid
 
 logger = logging.getLogger(__name__)
 
@@ -556,173 +551,3 @@ async def remove_all_projects_for_user(app: web.Application, user_id: int) -> No
                 new_project_owner_gid=new_project_owner_gid,
                 project=project,
             )
-
-
-async def get_new_project_owner_gid(
-    app: web.Application,
-    project_uuid: str,
-    user_id: int,
-    user_primary_gid: int,
-    project: Dict,
-) -> Optional[int]:
-    """Goes through the access rights and tries to find a new suitable owner.
-    The first viable user is selected as a new owner.
-    In order to become a new owner the user must have write access right.
-    """
-
-    access_rights = project["accessRights"]
-    # A Set[str] is prefered over Set[int] because access_writes
-    # is a Dict with only key,valus in {str, None}
-    other_users_access_rights: Set[str] = set(access_rights.keys()) - {
-        str(user_primary_gid)
-    }
-    logger.debug(
-        "Processing other user and groups access rights '%s'",
-        other_users_access_rights,
-    )
-
-    # Selecting a new project owner
-    # divide permissions between types of groups
-    standard_groups = {}  # groups of users, multiple users can be part of this
-    primary_groups = {}  # each individual user has a unique primary group
-    for other_gid in other_users_access_rights:
-        group: Optional[RowProxy] = await get_group_from_gid(
-            app=app, gid=int(other_gid)
-        )
-
-        # only process for users and groups with write access right
-        if group is None:
-            continue
-        if access_rights[other_gid]["write"] is not True:
-            continue
-
-        if group.type == GroupType.STANDARD:
-            standard_groups[other_gid] = access_rights[other_gid]
-        elif group.type == GroupType.PRIMARY:
-            primary_groups[other_gid] = access_rights[other_gid]
-
-    logger.debug(
-        "Possible new owner groups: standard='%s', primary='%s'",
-        standard_groups,
-        primary_groups,
-    )
-
-    new_project_owner_gid = None
-    # the primary group contains the users which which the project was directly shared
-    if len(primary_groups) > 0:
-        # fetch directly from the direct users with which the project is shared with
-        new_project_owner_gid = int(list(primary_groups.keys())[0])
-    # fallback to the groups search if the user does not exist
-    if len(standard_groups) > 0 and new_project_owner_gid is None:
-        new_project_owner_gid = await fetch_new_project_owner_from_groups(
-            app=app,
-            standard_groups=standard_groups,
-            user_id=user_id,
-        )
-
-    logger.info(
-        "Will move project '%s' to user with gid '%s', if user exists",
-        project_uuid,
-        new_project_owner_gid,
-    )
-
-    return new_project_owner_gid
-
-
-async def fetch_new_project_owner_from_groups(
-    app: web.Application, standard_groups: Dict, user_id: int
-) -> Optional[int]:
-    """Iterate over all the users in a group and if the users exists in the db
-    return its gid"""
-
-    # fetch all users in the group and then get their gid to put in here
-    # go through user_to_groups table and fetch all uid for matching gid
-    for group_gid in standard_groups.keys():
-        # remove the current owner from the bunch
-        target_group_users = await get_users_for_gid(app=app, gid=group_gid) - {user_id}
-        logger.error("Found group users '%s'", target_group_users)
-
-        for possible_user_id in target_group_users:
-            # check if the possible_user is still present in the db
-            try:
-                possible_user = await get_user(app=app, user_id=possible_user_id)
-                return int(possible_user["primary_gid"])
-            except users_exceptions.UserNotFoundError:
-                logger.warning(
-                    "Could not find new owner '%s' will try a new one",
-                    possible_user_id,
-                )
-
-        return None
-
-
-async def replace_current_owner(
-    app: web.Application,
-    project_uuid: str,
-    user_primary_gid: int,
-    new_project_owner_gid: int,
-    project: Dict,
-) -> None:
-    try:
-        new_project_owner_id = await get_user_id_from_gid(
-            app=app, primary_gid=new_project_owner_gid
-        )
-
-    except (
-        DatabaseError,
-        asyncpg.exceptions.PostgresError,
-        ProjectNotFoundError,
-        UserNotFoundError,
-    ):
-        logger.exception(
-            "Could not recover new user id from gid %s", new_project_owner_gid
-        )
-        return
-
-    # the result might me none
-    if new_project_owner_id is None:
-        logger.warning(
-            "Could not recover a new user id from gid %s", new_project_owner_gid
-        )
-        return
-
-    # unseting the project owner and saving the project back
-    project["prj_owner"] = int(new_project_owner_id)
-    # removing access rights entry
-    del project["accessRights"][str(user_primary_gid)]
-    project["accessRights"][
-        str(new_project_owner_gid)
-    ] = ProjectAccessRights.OWNER.value
-    logger.error("Syncing back project %s", project)
-
-    # syncing back project data
-    try:
-        await app[APP_PROJECT_DBAPI].update_project_without_checking_permissions(
-            project_data=project,
-            project_uuid=project_uuid,
-        )
-    except (
-        DatabaseError,
-        asyncpg.exceptions.PostgresError,
-        ProjectNotFoundError,
-        UserNotFoundError,
-    ):
-        logger.exception(
-            "Could not remove old owner and replaced it with user %s",
-            new_project_owner_id,
-        )
-
-
-async def remove_user(app: web.Application, user_id: int) -> None:
-    """Tries to remove a user, if the users still exists a warning message will be displayed"""
-    try:
-        await delete_user(app, user_id)
-    except (
-        DatabaseError,
-        asyncpg.exceptions.PostgresError,
-        ProjectNotFoundError,
-        UserNotFoundError,
-    ) as err:
-        logger.warning(
-            "User '%s' still has some projects, could not be deleted [%s]", user_id, err
-        )
